sys-apps/baselayout: Bump
Package-Manager: Portage-2.3.40, Repoman-2.3.9

+++ sys-apps/baselayout/baselayout-2.6-r1.ebuild
@@ -1,45 +1,72 @@
 # Copyright 1999-2018 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
 
-inherit eutils multilib
+EAPI=6
+
+inherit multilib versionator prefix
 
 DESCRIPTION="Filesystem baselayout and init scripts"
-HOMEPAGE="http://www.gentoo.org/"
-SRC_URI="mirror://gentoo/${P}.tar.bz2
-	http://dev.gentoo.org/~vapier/dist/${P}.tar.bz2"
+HOMEPAGE="https://www.gentoo.org/"
+if [[ ${PV} = 9999 ]]; then
+	EGIT_REPO_URI="https://anongit.gentoo.org/git/proj/${PN}.git"
+	inherit git-r3
+else
+	SRC_URI="https://gitweb.gentoo.org/proj/${PN}.git/snapshot/${P}.tar.bz2"
+	KEYWORDS="~alpha ~amd64 ~arm ~arm64 ~hppa ~ia64 ~m68k ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~x86 ~amd64-fbsd ~x86-fbsd"
+fi
 
 LICENSE="GPL-2"
 SLOT="0"
-KEYWORDS="alpha amd64 arm arm64 hppa ia64 m68k ~mips ppc ppc64 s390 sh sparc x86 ~amd64-fbsd ~x86-fbsd"
-IUSE="build kernel_linux"
+IUSE="build kernel_FreeBSD kernel_linux +split-usr"
 
 pkg_setup() {
 	multilib_layout
 }
 
 # Create our multilib dirs - the Makefile has no knowledge of this
-multilib_warn() {
-	local syms=$1 dirs=$2 def_libdir=$3
-
-	[ -z "${syms}${dirs}" ] && return
-
-	ewarn "Your system profile has SYMLINK_LIB=${SYMLINK_LIB}, so that means"
-	if [ -z "${syms}" ] ; then
-		ewarn "you need to have these paths as symlinks to ${def_libdir}:"
-		ewarn "$1"
-	fi
-}
 multilib_layout() {
-	local libdir libdirs=$(get_all_libdirs) def_libdir=$(get_abi_LIBDIR $DEFAULT_ABI)
+	local dir def_libdir libdir libdirs
+	local prefix prefix_lst
+	def_libdir=$(get_abi_LIBDIR $DEFAULT_ABI)
+	libdirs=$(get_all_libdirs)
 	: ${libdirs:=lib}	# it isn't that we don't trust multilib.eclass...
 
-	[ -z "${def_libdir}" ] && die "your DEFAULT_ABI=$DEFAULT_ABI appears to be invalid"
+	if [[ -z "${SYMLINK_LIB}" || ${SYMLINK_LIB} = no ]] ; then
+		prefix_lst=( "${EROOT}"{,usr/,usr/local/} )
+		for prefix in ${prefix_lst[@]}; do
+			for libdir in ${libdirs}; do
+				dir="${prefix}${libdir}"
+				if [[ -e "${dir}" ]]; then
+					[[ ! -d "${dir}" ]] &&
+						die "${dir} exists but is not a directory"
+					continue
+				fi
+				if ! use split-usr && [[ ${prefix} = ${EROOT} ]]; then
+					einfo "symlinking ${dir} to usr/${libdir}"
+					ln -s usr/${libdir} ${dir} ||
+						die " Unable to make ${dir} symlink"
+				else
+					einfo "creating directory ${dir}"
+					mkdir -p "${dir}" ||
+						die "Unable to create ${dir} directory"
+				fi
+			done
+		done
+		return 0
+	fi
+
+	[ -z "${def_libdir}" ] &&
+		die "your DEFAULT_ABI=$DEFAULT_ABI appears to be invalid"
 
 	# figure out which paths should be symlinks and which should be directories
 	local dirs syms exp d
 	for libdir in ${libdirs} ; do
-		exp=( {,usr/,usr/local/}${libdir} )
-		for d in "${exp[@]/#/${ROOT}}" ; do
+		if use split-usr ; then
+			exp=( {,usr/,usr/local/}${libdir} )
+		else
+			exp=( {usr/,usr/local/}${libdir} )
+		fi
+		for d in "${exp[@]}" ; do
 			# most things should be dirs
 			if [ "${SYMLINK_LIB}" = "yes" ] && [ "${libdir}" = "lib" ] ; then
 				[ ! -h "${d}" ] && [ -e "${d}" ] && dirs+=" ${d}"
@@ -60,8 +87,12 @@ multilib_layout() {
 
 	# setup symlinks and dirs where we expect them to be; do not migrate
 	# data ... just fall over in that case.
-	local prefix
-	for prefix in "${ROOT}"{,usr/,usr/local/} ; do
+	if use split-usr ; then
+		prefix_lst=( "${EROOT}"{,usr/,usr/local/} )
+	else
+		prefix_lst=( "${EROOT}"{usr/,usr/local/} )
+	fi
+	for prefix in "${prefix_lst[@]}"; do
 		if [ "${SYMLINK_LIB}" = yes ] ; then
 			# we need to make sure "lib" points to the native libdir
 			if [ -h "${prefix}lib" ] ; then
@@ -96,14 +127,14 @@ multilib_layout() {
 				else
 					mkdir -p "${prefix}lib" || die
 				fi
-			elif [ -d "${prefix}lib" ] ; then
+			elif [ -d "${prefix}lib" ] && ! has lib32 ${libdirs} ; then
 				# make sure the old "lib" ABI location does not exist; we
 				# only symlinked the lib dir on systems where we moved it
 				# to "lib32" ...
 				case ${CHOST} in
 				*-gentoo-freebsd*) ;; # We want it the other way on fbsd.
 				i?86*|x86_64*|powerpc*|sparc*|s390*)
-					if [ -d "${prefix}lib32" ] ; then
+					if [[ -d ${prefix}lib32 && ! -h ${prefix}lib32 ]] ; then
 						rm -f "${prefix}lib32"/.keep
 						if ! rmdir "${prefix}lib32" 2>/dev/null ; then
 							ewarn "You need to merge ${prefix}lib32 into ${prefix}lib"
@@ -115,35 +146,37 @@ multilib_layout() {
 			else
 				# nothing exists, so just set it up sanely
 				ewarn "Initializing ${prefix}lib as a dir"
-				mkdir -p "${prefix}" || die
-				rm -f "${prefix}lib" || die
-				ln -s ${def_libdir} "${prefix}lib" || die
+				mkdir -p "${prefix}lib" || die
 			fi
 		fi
 	done
+	if ! use split-usr ; then
+		for libdir in ${libdirs}; do
+			if [[ ! -e "${EROOT}${libdir}" ]]; then
+				ln -s usr/"${libdir}" "${EROOT}${libdir}"
+			fi
+		done
+	fi
 }
 
 pkg_preinst() {
-	# Bug #217848 - Since the remap_dns_vars() called by pkg_preinst() of
-	# the baselayout-1.x ebuild copies all the real configs from the user's
-	# /etc/conf.d into ${D}, it makes them all appear to be the default
-	# versions. In order to protect them from being unmerged after this
-	# upgrade, modify their timestamps.
-	touch "${ROOT}"/etc/conf.d/* 2>/dev/null
-
 	# This is written in src_install (so it's in CONTENTS), but punt all
 	# pending updates to avoid user having to do etc-update (and make the
 	# pkg_postinst logic simpler).
-	rm -f "${ROOT}"/etc/._cfg????_gentoo-release
+	rm -f "${EROOT}"/etc/._cfg????_gentoo-release
 
 	# We need to install directories and maybe some dev nodes when building
 	# stages, but they cannot be in CONTENTS.
 	# Also, we cannot reference $S as binpkg will break so we do this.
 	multilib_layout
 	if use build ; then
-		emake -C "${D}/usr/share/${PN}" DESTDIR="${ROOT}" layout || die
+		if use split-usr ; then
+			emake -C "${ED}/usr/share/${PN}" DESTDIR="${EROOT}" layout
+		else
+			emake -C "${ED}/usr/share/${PN}" DESTDIR="${EROOT}" layout-usrmerge
+		fi
 	fi
-	rm -f "${D}"/usr/share/${PN}/Makefile
+	rm -f "${ED}"/usr/share/${PN}/Makefile
 
 	# Sabayon customization, protect /etc/hosts from removal (from older ebuilds)
 	local etc_hosts="${ROOT}/etc/hosts"
@@ -157,22 +190,17 @@ src_unpack() {
 
 	cd "${S}"
 	# We are Sabayon!
-	epatch "${FILESDIR}/${PN}-sabayon-os-release.patch"
+	eapply "${FILESDIR}/${PN}-sabayon-os-release.patch"
 }
 
-src_install() {
-	# Setup /run directory, this is missing from original baselayout
-	dodir /run
-
-	emake \
-		OS=$(usex kernel_FreeBSD BSD Linux) \
-		DESTDIR="${D}" \
-		install || die
-	dodoc ChangeLog.svn
-
-	# need the makefile in pkg_preinst
-	insinto /usr/share/${PN}
-	doins Makefile || die
+src_prepare() {
+	default
+	if use prefix; then
+		hprefixify -e "/EUID/s,0,${EUID}," -q '"' etc/profile
+		hprefixify etc/{env.d/50baselayout,shells} share.Linux/passwd
+		echo PATH=/usr/bin:/bin >> etc/env.d/99host
+		echo ROOTPATH=/usr/sbin:/sbin:/usr/bin:/bin >> etc/env.d/99host
+	fi
 
 	# handle multilib paths.  do it here because we want this behavior
 	# regardless of the C library that you're using.  we do explicitly
@@ -182,16 +210,25 @@ src_install() {
 	# path and the symlinked path doesn't change the resulting cache.
 	local libdir ldpaths
 	for libdir in $(get_all_libdirs) ; do
-		ldpaths+=":/${libdir}:/usr/${libdir}:/usr/local/${libdir}"
+		ldpaths+=":${EPREFIX}/${libdir}:${EPREFIX}/usr/${libdir}"
+		ldpaths+=":${EPREFIX}/usr/local/${libdir}"
 	done
-	echo "LDPATH='${ldpaths#:}'" >> "${D}"/etc/env.d/00basic
+	echo "LDPATH='${ldpaths#:}'" >> etc/env.d/50baselayout
 
 	# rc-scripts version for testing of features that *should* be present
-	echo "Gentoo Base System release ${PV}" > "${D}"/etc/gentoo-release
+	echo "Gentoo Base System release ${PV}" > etc/gentoo-release
+}
 
-	# Sabayon customization, install /etc/hosts separately (to .example)
-	mv "${D}"/etc/hosts "${D}"/etc/hosts.example || die "cannot move /etc/hosts"
+src_install() {
+	emake \
+		OS=$(usex kernel_FreeBSD BSD Linux) \
+		DESTDIR="${ED}" \
+		install
+	dodoc ChangeLog
 
+	# need the makefile in pkg_preinst
+	insinto /usr/share/${PN}
+	doins Makefile
 }
 
 pkg_postinst() {
@@ -203,24 +240,24 @@ pkg_postinst() {
 	# (3) accidentally packaging up personal files with quickpkg
 	# If they don't exist then we install them
 	for x in master.passwd passwd shadow group fstab ; do
-		[ -e "${ROOT}etc/${x}" ] && continue
-		[ -e "${ROOT}usr/share/baselayout/${x}" ] || continue
-		cp -p "${ROOT}usr/share/baselayout/${x}" "${ROOT}"etc
+		[ -e "${EROOT}etc/${x}" ] && continue
+		[ -e "${EROOT}usr/share/baselayout/${x}" ] || continue
+		cp -p "${EROOT}usr/share/baselayout/${x}" "${EROOT}"etc
 	done
 
 	# Force shadow permissions to not be world-readable #260993
 	for x in shadow ; do
-		[ -e "${ROOT}etc/${x}" ] && chmod o-rwx "${ROOT}etc/${x}"
+		[ -e "${EROOT}etc/${x}" ] && chmod o-rwx "${EROOT}etc/${x}"
 	done
 
 	# Take care of the etc-update for the user
-	if [ -e "${ROOT}"/etc/._cfg0000_gentoo-release ] ; then
-		mv "${ROOT}"/etc/._cfg0000_gentoo-release "${ROOT}"/etc/gentoo-release
+	if [ -e "${EROOT}"etc/._cfg0000_gentoo-release ] ; then
+		mv "${EROOT}"etc/._cfg0000_gentoo-release "${EROOT}"etc/gentoo-release
 	fi
 
 	# whine about users that lack passwords #193541
-	if [[ -e ${ROOT}/etc/shadow ]] ; then
-		local bad_users=$(sed -n '/^[^:]*::/s|^\([^:]*\)::.*|\1|p' "${ROOT}"/etc/shadow)
+	if [[ -e "${EROOT}"etc/shadow ]] ; then
+		local bad_users=$(sed -n '/^[^:]*::/s|^\([^:]*\)::.*|\1|p' "${EROOT}"/etc/shadow)
 		if [[ -n ${bad_users} ]] ; then
 			echo
 			ewarn "The following users lack passwords!"
@@ -228,13 +265,9 @@ pkg_postinst() {
 		fi
 	fi
 
-	# baselayout leaves behind a lot of .keep files, so let's clean them up
-	find "${ROOT}"/lib*/rcscripts/ -name .keep -exec rm -f {} + 2>/dev/null
-	find "${ROOT}"/lib*/rcscripts/ -depth -type d -exec rmdir {} + 2>/dev/null
-
 	# whine about users with invalid shells #215698
-	if [[ -e ${ROOT}/etc/passwd ]] ; then
-		local bad_shells=$(awk -F: 'system("test -e " $7) { print $1 " - " $7}' /etc/passwd | sort)
+	if [[ -e "${EROOT}"etc/passwd ]] ; then
+		local bad_shells=$(awk -F: 'system("test -e " $7) { print $1 " - " $7}' "${EROOT}"etc/passwd | sort)
 		if [[ -n ${bad_shells} ]] ; then
 			echo
 			ewarn "The following users have non-existent shells!"
@@ -242,22 +275,37 @@ pkg_postinst() {
 		fi
 	fi
 
-	# Sabayon customization, copy /etc/hosts back in place if it doesn't exist
-	local etc_hosts="${ROOT}/etc/hosts"
-	if [ -e "${etc_hosts}.baselayout_ebuild_backup" ]; then
-		cp -p "${etc_hosts}.baselayout_ebuild_backup" "${etc_hosts}" # don't die
-	elif [ ! -e "${etc_hosts}" ]; then
-		cp -p "${etc_hosts}.example" "${etc_hosts}" # don't die
-	fi
-	chown root:root "${etc_hosts}" # don't die
-
-	# http://bugs.gentoo.org/361349
+	# https://bugs.gentoo.org/361349
 	if use kernel_linux; then
-		mkdir -p "${ROOT}"/run
+		mkdir -p "${EROOT}"run
 
-		if ! grep -qs "^tmpfs.*/run " "${ROOT}"/proc/mounts ; then
-			echo
-			ewarn "You should reboot the system now to get /run mounted with tmpfs!"
+		local found fstype mountpoint
+		while read -r _ mountpoint fstype _; do
+		[[ ${mountpoint} = /run ]] && [[ ${fstype} = tmpfs ]] && found=1
+		done < "${ROOT}"proc/mounts
+		[[ -z ${found} ]] &&
+			ewarn "You should reboot now to get /run mounted with tmpfs!"
+	fi
+
+	for x in ${REPLACING_VERSIONS}; do
+		if ! version_is_at_least 2.4 ${x}; then
+			ewarn "After updating ${EROOT}etc/profile, please run"
+			ewarn "env-update && . /etc/profile"
 		fi
+
+		if ! version_is_at_least 2.6 ${x}; then
+			ewarn "Please run env-update then log out and back in to"
+			ewarn "update your path."
+		fi
+		# clean up after 2.5 typos
+		# https://bugs.gentoo.org/show_bug.cgi?id=656380
+		if [[ ${x} == 2.5 ]]; then
+			rm -fr "${EROOT}{,usr"
+		fi
+	done
+
+	if [[ -e "${EROOT}"etc/env.d/00basic ]]; then
+		ewarn "${EROOT}etc/env.d/00basic is now ${EROOT}etc/env.d/50baselayout"
+		ewarn "Please migrate your changes."
 	fi
 }
